LayoutTests/fast/css-grid-layout/non-grid-columns-rows-get-set-multiple.html

<!DOCTYPE html>
<html>
<head>
<script>
if (window.testRunner)
    testRunner.overridePreference("WebKitCSSGridLayoutEnabled", 1);
</script>
<link href="resources/grid.css" rel="stylesheet">
<style>
.gridWithFixed {
    -webkit-grid-template-columns: 7px 11px;
    -webkit-grid-template-rows: 17px 2px;
}

.gridWithPercent {
    -webkit-grid-template-columns: 53% 99%;
    -webkit-grid-template-rows: 27% 52%;
}
.gridWithAuto {
    -webkit-grid-template-columns: auto auto;
    -webkit-grid-template-rows: auto auto;
}
.gridWithEM {
    -webkit-grid-template-columns: 10em 12em;
    -webkit-grid-template-rows: 15em 17em;
    font: 10px Ahem;
}
.gridWithNoneAndAuto {
    -webkit-grid-template-columns: none auto;
    -webkit-grid-template-rows: none auto;
}
.gridNoneWithAndFixed {
    -webkit-grid-template-columns: none 15px;
    -webkit-grid-template-rows: none 22px;
}
.gridWithThreeItems {
    -webkit-grid-template-columns: 15px auto 10em;
    -webkit-grid-template-rows: 12em 18px auto;
    font: 10px Ahem;
}
.gridWithPercentAndViewportPercent {
    -webkit-grid-template-columns: 50% 15vw;
    -webkit-grid-template-rows: 35% 28vh;
}
.gridWithFitContentAndFitAvailable {
    -webkit-grid-template-columns: -webkit-content-available;
    -webkit-grid-template-rows: -webkit-fit-content -webkit-fit-available;
}
.gridWithMinMaxContent {
    -webkit-grid-template-columns: -webkit-min-content -webkit-max-content;
    -webkit-grid-template-rows: -webkit-max-content -webkit-min-content;
}
.gridWithMinMaxAndFixed {
    -webkit-grid-template-columns: minmax(45px, 30%) 15px;
    -webkit-grid-template-rows: 12em minmax(35%, 10px);
    font: 10px Ahem;
}
.gridWithMinMaxAndMinMaxContent {
    -webkit-grid-template-columns: minmax(-webkit-min-content, 30%) 15px;
    -webkit-grid-template-rows: 12em minmax(35%, -webkit-max-content);
    font: 10px Ahem;
}
.gridWithFractionFraction {
    -webkit-grid-template-columns: 1fr 2fr;
    -webkit-grid-template-rows: 3fr 4fr;
}
.gridWithFractionMinMax {
    -webkit-grid-template-columns: minmax(-webkit-min-content, 45px) 2fr;
    -webkit-grid-template-rows: 3fr minmax(14px, -webkit-max-content);
}
</style>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<div class="gridWithFixed" id="gridWithFixedElement"></div>
<div class="gridWithPercent" id="gridWithPercentElement"></div>
<div class="gridWithAuto" id="gridWithAutoElement"></div>
<div class="gridWithEM" id="gridWithEMElement"></div>
<div class="gridWithNoneAndAuto" id="gridWithNoneAndAuto"></div>
<div class="gridWithNoneAndFixed" id="gridWithNoneAndFixed"></div>
<div class="gridWithThreeItems" id="gridWithThreeItems"></div>
<div class="gridWithPercentAndViewportPercent" id="gridWithPercentAndViewportPercent"></div>
<div class="gridWithFitContentAndFitAvailable" id="gridWithFitContentAndFitAvailable"></div>
<div class="gridWithMinMaxContent" id="gridWithMinMaxContent"></div>
<div class="gridWithMinMaxAndFixed" id="gridWithMinMaxAndFixed"></div>
<div class="gridWithMinMaxAndMinMaxContent" id="gridWithMinMaxAndMinMaxContent"></div>
<div class="gridWithFractionFraction" id="gridWithFractionFraction"></div>
<div class="gridWithFractionMinMax" id="gridWithFractionMinMax"></div>
<script src="resources/grid-definitions-parsing-utils.js"></script>
<script src="resources/non-grid-columns-rows-get-set-multiple.js"></script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
